requirements.txt
# Host-side Python (host_control uart/execute, runtime_collector)
pyserial>=3.5
